Added log function.
TODO:  Right now, when the Arduino boots up, it prints the initialization values from the loop, but then nothing happens.  I think it is an issue with initializing the CAES object.  We should put debugging statements in each of the separate classes, so we know exactly what happens.  More information about the debugging levels is in config.h.