Changes to logging and PID reset/stop/run calls
I simplified the logging and in the process discovered that logging statements take up 2% of memory (regardless of debug level). Perhaps if we build the debug level check and serial printing in different functions that would remove the unnecessary memory. I'm not sure what's going on behind the scenes.

I also changed how PID run stop and reset are being used to conform to the documentation, which I added in comments next to the calls. This needs to be tested. We may need to dig into the source code for AutoPID to verify that it is doing what we expect before we kill ourselves debugging.